python/requirements.txt
PyYAML==3.11
argparse==1.2.1
pep8==1.7.0
pprintpp==0.2.3
pyOpenSSL==16.0.0
requests==2.10.0
six==1.10.0
urllib3==1.16
wsgiref==0.1.2
